requirements.txt
pymongo==4.8.0
fastapi==0.115.0  # Nếu dùng FastAPI
uvicorn==0.30.6   # Server cho FastAPI
pyjwt==2.9.0
python-multipart==0.0.12